no red tests

+++ pallets/subtensor/tests/epoch.rs
@@ -1911,7 +1911,7 @@ fn test_validator_permits() {
                         SubtensorModule::increase_stake_on_coldkey_hotkey_account(
                             &U256::from(key),
                             &U256::from(key),
-                            network_n,
+                            network_n as u16,
                             stake[key as usize],
                         );
                     }
@@ -1946,7 +1946,7 @@ fn test_validator_permits() {
                         SubtensorModule::increase_stake_on_coldkey_hotkey_account(
                             &(U256::from(*server as u64)),
                             &(U256::from(*server as u64)),
-                            network_n,
+                            network_n as u16,
                             2 * network_n as u64,
                         );
                     }
